refactor(admin): share target runtime status collection
